ts clean up
