+++ src/components/form/form_row/form_row.test.tsx
@@ -17,20 +17,6 @@ describe('EuiFormRow', () => {
     expect(component).toMatchSnapshot();
   });
 
-  test('no children is an error', () => {
-    // @ts-ignore
-    expect(() => <EuiFormRow {...requiredProps} />).toThrow();
-  });
-
-  test('two children is an error', () => {
-    expect(() => (
-      <EuiFormRow {...requiredProps}>
-        <div />
-        <div />
-      </EuiFormRow>
-    )).toThrow();
-  });
-
   test('ties together parts for accessibility', () => {
     const props = {
       label: 'Label',
