--- src/components/form/form_row/index.d.ts
@@ -1,51 +1,0 @@
-import { CommonProps, ExclusiveUnion } from '../../common';
-
-import {
-  FunctionComponent,
-  ReactNode,
-  ReactElement,
-  HTMLAttributes,
-} from 'react';
-
-declare module '@elastic/eui' {
-  /**
-   * @see './form_row.ts'
-   */
-  export type EuiFormRowCommonProps = CommonProps & {
-    error?: ReactNode | ReactNode[];
-    fullWidth?: boolean;
-    hasEmptyLabelSpace?: boolean;
-    helpText?: ReactNode;
-    isInvalid?: boolean;
-    label?: ReactNode;
-    labelAppend?: ReactNode;
-    describedByIds?: string[];
-    display?:
-      | 'row'
-      | 'rowCompressed'
-      | 'columnCompressed'
-      | 'center'
-      | 'centerCompressed'
-      | 'columnCompressedSwitch';
-    // **DEPRECATED: use `display: rowCompressed` instead.**
-    compressed?: boolean;
-    // **DEPRECATED: use `display: center` instead.**
-    displayOnly?: boolean;
-  };
-
-  type LabelProps = {
-    labelType?: 'label';
-  } & EuiFormRowCommonProps &
-    HTMLAttributes<HTMLDivElement>;
-
-  type LegendProps = {
-    labelType?: 'legend';
-  } & EuiFormRowCommonProps &
-    HTMLAttributes<HTMLFieldSetElement>;
-
-  export type EuiFormRowProps = ExclusiveUnion<LabelProps, LegendProps> & {
-    children: ReactElement;
-  };
-
-  export const EuiFormRow: FunctionComponent<EuiFormRowProps>;
-}
